services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# bcrypt 4.1+ quebra a detecção de backend do passlib (ValueError nos 72 bytes no registo).
bcrypt>=4.0.1,<4.1.0
python-multipart>=0.0.12
pydantic-settings>=2.6.0
redis>=5.0.0
docker>=7.0.0
grpcio>=1.62.0
grpcio-tools>=1.62.0
pydantic>=2.0.0
email-validator>=2.0.0
